requirements.txt
pandas>=1.3.0
openpyxl>=3.0.9
xlrd>=2.0.1
python-docx>=0.8.11

